Moved non-local means denoising to restoration submodule
