+++ skimage/filters/setup.py
@@ -14,7 +14,6 @@ def configuration(parent_package='', top_path=None):
     config.add_data_dir('rank/tests')
 
     cython(['_ctmf.pyx'], working_path=base_path)
-    cython(['_nl_means_denoising.pyx'], working_path=base_path)
     cython(['rank/core_cy.pyx'], working_path=base_path)
     cython(['rank/generic_cy.pyx'], working_path=base_path)
     cython(['rank/percentile_cy.pyx'], working_path=base_path)
@@ -22,9 +21,6 @@ def configuration(parent_package='', top_path=None):
 
     config.add_extension('_ctmf', sources=['_ctmf.c'],
                          include_dirs=[get_numpy_include_dirs()])
-    config.add_extension('_nl_means_denoising',
-        sources=['_nl_means_denoising.c'],
-        include_dirs=[get_numpy_include_dirs()])
     config.add_extension('rank.core_cy', sources=['rank/core_cy.c'],
         include_dirs=[get_numpy_include_dirs()])
     config.add_extension('rank.generic_cy', sources=['rank/generic_cy.c'],
