+++ skimage/restoration/__init__.py
@@ -22,7 +22,7 @@ from .deconvolution import wiener, unsupervised_wiener, richardson_lucy
 from .unwrap import unwrap_phase
 from ._denoise import denoise_tv_chambolle, denoise_tv_bregman, \
                       denoise_bilateral
-
+from .nl_means_denoising import nl_means_denoising
 
 __all__ = ['wiener',
            'unsupervised_wiener',
@@ -30,4 +30,5 @@ __all__ = ['wiener',
            'unwrap_phase',
            'denoise_tv_bregman',
            'denoise_tv_chambolle',
-           'denoise_bilateral']
+           'denoise_bilateral',
+           'nl_means_denoising']
