+++ skimage/restoration/_nl_means_denoising.pyx
