+++ skimage/restoration/nl_means_denoising.py
