+++ skimage/restoration/setup.py
@@ -17,6 +17,7 @@ def configuration(parent_package='', top_path=None):
     cython(['_unwrap_2d.pyx'], working_path=base_path)
     cython(['_unwrap_3d.pyx'], working_path=base_path)
     cython(['_denoise_cy.pyx'], working_path=base_path)
+    cython(['_nl_means_denoising.pyx'], working_path=base_path)
 
     config.add_extension('_unwrap_1d', sources=['_unwrap_1d.c'],
                          include_dirs=[get_numpy_include_dirs()])
@@ -28,6 +29,9 @@ def configuration(parent_package='', top_path=None):
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_denoise_cy', sources=['_denoise_cy.c'],
         include_dirs=[get_numpy_include_dirs(), '../_shared'])
+    config.add_extension('_nl_means_denoising',
+                         sources=['_nl_means_denoising.c'],
+                         include_dirs=[get_numpy_include_dirs()])
 
     return config
